#4 Integrate evaluation script

+++ src/main/kotlin/com/akarakoutev/agmdoz/core/ChatMessage.kt
@@ -1,7 +1,6 @@
 package com.akarakoutev.agmdoz.core
 
 import java.time.Instant
-import java.util.UUID
 import javax.persistence.Entity
 import javax.persistence.GeneratedValue
 import javax.persistence.Id
@@ -14,7 +13,7 @@ data class ChatMessage(
     var modelVersion: String?,
     val ts: Instant = Instant.now(),
     val text: String,
-    var type: Pair<MessageType, Float>?)
+    var type: Pair<MessageType, Double>?)
 
 enum class MessageType {
     POSITIVE,
